backend/requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2
google-generativeai==0.8.3
supabase==2.7.1
python-dotenv==1.0.1
requests==2.32.3